wordofra/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-pytest==0.2.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
